requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-aarch64
# created-by: conda 24.11.3
_openmp_mutex=4.5=2_gnu
accelerate=1.3.0=pypi_0
aiofiles=23.2.1=pypi_0
aiohappyeyeballs=2.4.4=pypi_0
aiohttp=3.11.11=pypi_0
aiosignal=1.3.2=pypi_0
altair=5.5.0=pypi_0
annotated-types=0.7.0=pypi_0
anthropic=0.45.0=pypi_0
anyio=4.8.0=pypi_0
asttokens=3.0.0=pyhd8ed1ab_1
async-timeout=5.0.1=pypi_0
attrs=25.1.0=pypi_0
azure-ai-inference=1.0.0b9=pypi_0
azure-core=1.32.0=pypi_0
bzip2=1.0.8=h68df207_7
ca-certificates=2025.1.31=hcefe29a_0
certifi=2024.12.14=pypi_0
charset-normalizer=3.4.1=pypi_0
click=8.1.8=pypi_0
cloudpickle=3.1.1=pypi_0
comm=0.2.2=pyhd8ed1ab_1
contourpy=1.3.1=pypi_0
cycler=0.12.1=pypi_0
datasets=3.2.0=pypi_0
debugpy=1.8.12=py310he30c3ed_0
decorator=5.1.1=pyhd8ed1ab_1
dill=0.3.8=pypi_0
distro=1.9.0=pypi_0
docker-pycreds=0.4.0=pypi_0
einops=0.8.0=pypi_0
eval-type-backport=0.2.2=pypi_0
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.1.0=pyhd8ed1ab_1
farama-notifications=0.0.4=pypi_0
fastapi=0.115.7=pypi_0
fastchat=0.1.0=pypi_0
ffmpy=0.5.0=pypi_0
filelock=3.16.1=pypi_0
fonttools=4.55.6=pypi_0
frozenlist=1.5.0=pypi_0
fschat=0.2.36=pypi_0
fsspec=2024.10.0=pypi_0
git-filter-repo=2.47.0=pypi_0
gitdb=4.0.12=pypi_0
gitpython=3.1.44=pypi_0
gradio=3.50.2=pypi_0
gradio-client=0.6.1=pypi_0
gymnasium=1.1.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.7=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.30.2=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.6.1=pyha770c72_0
importlib-resources=6.5.2=pypi_0
ipykernel=6.29.5=pyh3099207_0
ipython=8.32.0=pyh907856f_0
isodate=0.7.2=pypi_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.4=pypi_0
jiter=0.8.2=pypi_0
joblib=1.4.2=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=pyh31011fe_1
keyutils=1.6.1=h4e544f5_0
kiwisolver=1.4.8=pypi_0
krb5=1.21.3=h50a48e9_0
latex2mathml=3.77.0=pypi_0
ld_impl_linux-aarch64=2.43=h80caac9_2
libedit=3.1.20250104=pl5321h976ea20_0
libffi=3.4.2=h3557bc0_5
libgcc=14.2.0=he277a41_1
libgcc-ng=14.2.0=he9431aa_1
libgomp=14.2.0=he277a41_1
liblzma=5.6.3=h86ecc28_1
libnsl=2.0.1=h31becfc_0
libsodium=1.0.20=h68df207_0
libsqlite=3.48.0=h5eb1b54_1
libstdcxx=14.2.0=h3f4de04_1
libstdcxx-ng=14.2.0=hf1166c9_1
libuuid=2.38.1=hb4cce97_0
libxcrypt=4.4.36=h31becfc_1
libzlib=1.3.1=h86ecc28_2
markdown-it-py=3.0.0=pypi_0
markdown2=2.5.3=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.10.0=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mdurl=0.1.2=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.1.0=pypi_0
multiprocess=0.70.16=pypi_0
narwhals=1.24.1=pypi_0
ncurses=6.5=ha32ae93_2
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.4.2=pypi_0
nh3=0.2.20=pypi_0
nltk=3.9.1=pypi_0
numpy=2.2.2=pypi_0
nvidia-ml-py=12.570.86=pypi_0
openai=1.60.1=pypi_0
openssl=3.4.0=hd08dc88_1
orjson=3.10.15=pypi_0
packaging=24.2=pyhd8ed1ab_2
pandas=2.2.3=pypi_0
parso=0.8.4=pyhd8ed1ab_1
peft=0.14.0=pypi_0
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.1.0=pypi_0
pip=24.3.1=pyh8b19718_2
platformdirs=4.3.6=pyhd8ed1ab_1
prompt-toolkit=3.0.50=pyha770c72_0
propcache=0.2.1=pypi_0
protobuf=5.29.3=pypi_0
psutil=6.1.1=py310ha766c32_0
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pyarrow=19.0.0=pypi_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pydub=0.25.1=pypi_0
pygame=2.6.1=pypi_0
pygments=2.19.1=pyhd8ed1ab_0
pynvml=12.0.0=pypi_0
pyparsing=3.2.1=pypi_0
python=3.10.16=h57b00e1_1_cpython
python-dateutil=2.9.0.post0=pyhff2d567_1
python-dotenv=1.0.1=pypi_0
python-multipart=0.0.20=pypi_0
python_abi=3.10=5_cp310
pytz=2024.2=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=26.2.1=py310h55e1596_0
readline=8.2=h8fc344f_1
referencing=0.36.2=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
rich=13.9.4=pypi_0
rpds-py=0.22.3=pypi_0
safetensors=0.5.2=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.15.1=pypi_0
semantic-version=2.10.0=pypi_0
sentencepiece=0.2.0=pypi_0
sentry-sdk=2.20.0=pypi_0
setproctitle=1.3.4=pypi_0
setuptools=75.8.0=pyhff2d567_0
shellingham=1.5.4=pypi_0
shortuuid=1.0.13=pypi_0
six=1.17.0=pyhd8ed1ab_0
smmap=5.0.2=pypi_0
sniffio=1.3.1=pypi_0
stack_data=0.6.3=pyhd8ed1ab_1
starlette=0.45.3=pypi_0
svgwrite=1.4.3=pypi_0
sympy=1.13.1=pypi_0
tabulate=0.9.0=pypi_0
threadpoolctl=3.6.0=pypi_0
tiktoken=0.8.0=pypi_0
tk=8.6.13=h194ca79_0
together=1.4.1=pypi_0
tokenizers=0.15.2=pypi_0
torch=2.6.0.dev20241210+cu124=pypi_0
torchaudio=2.6.0.dev20241216=pypi_0
torchvision=0.22.0.dev20241216=pypi_0
tornado=6.4.2=py310h78583b1_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
transformers=4.38.2=pypi_0
transformers-stream-generator=0.0.5=pypi_0
typer=0.15.1=pypi_0
typing_extensions=4.12.2=pyha770c72_1
tzdata=2025.1=pypi_0
urllib3=2.3.0=pypi_0
uvicorn=0.34.0=pypi_0
wandb=0.18.7=pypi_0
wavedrom=2.0.3.post3=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_1
websockets=11.0.3=pypi_0
wheel=0.45.1=pyhd8ed1ab_1
xxhash=3.5.0=pypi_0
yarl=1.18.3=pypi_0
zeromq=4.3.5=h5efb499_7
zipp=3.21.0=pyhd8ed1ab_1
